requirements.txt
matplotlib==3.8.0
pandas==2.0.3
pytrends==4.9.2
python-dotenv==1.0.0